+++ docs/pages/store/_meta.js
@@ -7,5 +7,5 @@ export default {
   spec: "Specification",
   "gas-efficiency": "Gas efficiency",
   "using-without-world": "Using Store without World",
-  "internals": "Internal tables",
+  internals: "Internal tables",
 };