+++ examples/minimal/packages/client-vanilla/src/mud/createSystemCalls.ts
@@ -1,18 +1,18 @@
 import { getComponentValue } from "@latticexyz/recs";
-import { awaitStreamValue } from "@latticexyz/utils";
 import { ClientComponents } from "./createClientComponents";
 import { SetupNetworkResult } from "./setupNetwork";
+import { singletonEntity } from "@latticexyz/store-sync/recs";
 
 export type SystemCalls = ReturnType<typeof createSystemCalls>;
 
 export function createSystemCalls(
-  { worldSend, txReduced$, singletonEntity }: SetupNetworkResult,
-  { Counter }: ClientComponents
+  { worldContract, waitForTransaction }: SetupNetworkResult,
+  { CounterTable }: ClientComponents
 ) {
   const increment = async () => {
-    const tx = await worldSend("increment", []);
-    await awaitStreamValue(txReduced$, (txHash) => txHash === tx.hash);
-    return getComponentValue(Counter, singletonEntity);
+    const tx = await worldContract.write.increment();
+    await waitForTransaction(tx);
+    return getComponentValue(CounterTable, singletonEntity);
   };
 
   return {