+++ packages/store/test/Schema.t.sol
@@ -5,12 +5,13 @@ import { Test, console } from "forge-std/Test.sol";
 import { GasReporter } from "@latticexyz/gas-report/src/GasReporter.sol";
 import { SchemaType } from "@latticexyz/schema-type/src/solidity/SchemaType.sol";
 import { Schema, SchemaLib } from "../src/Schema.sol";
+import { SchemaEncodeHelper } from "./SchemaEncodeHelper.sol";
 
 // TODO add tests for all schema types
 contract SchemaTest is Test, GasReporter {
   function testEncodeDecodeSchema() public {
-    startGasReport("encode schema with 6 entries [SchemaLib.encode]");
-    Schema schema = SchemaLib.encode(
+    startGasReport("encode schema with 6 entries");
+    Schema schema = SchemaEncodeHelper.encode(
       SchemaType.UINT8, // 1 byte
       SchemaType.UINT16, // 2 bytes
       SchemaType.UINT32, // 4 bytes
@@ -33,7 +34,7 @@ contract SchemaTest is Test, GasReporter {
   }
 
   function testFailInvalidSchemaStaticAfterDynamic() public pure {
-    SchemaLib.encode(SchemaType.UINT8, SchemaType.UINT32_ARRAY, SchemaType.UINT16);
+    SchemaEncodeHelper.encode(SchemaType.UINT8, SchemaType.UINT32_ARRAY, SchemaType.UINT16);
   }
 
   function testEncodeMaxValidLength() public {
@@ -129,7 +130,7 @@ contract SchemaTest is Test, GasReporter {
   }
 
   function testGetStaticSchemaLength() public {
-    Schema schema = SchemaLib.encode(
+    Schema schema = SchemaEncodeHelper.encode(
       SchemaType.UINT8, // 1 byte
       SchemaType.UINT16, // 2 bytes
       SchemaType.UINT32, // 4 bytes
@@ -146,7 +147,7 @@ contract SchemaTest is Test, GasReporter {
   }
 
   function testGetNumStaticFields() public {
-    Schema schema = SchemaLib.encode(
+    Schema schema = SchemaEncodeHelper.encode(
       SchemaType.UINT8, // 1 byte
       SchemaType.UINT16, // 2 bytes
       SchemaType.UINT32, // 4 bytes
@@ -163,7 +164,7 @@ contract SchemaTest is Test, GasReporter {
   }
 
   function testGetNumDynamicFields() public {
-    Schema schema = SchemaLib.encode(
+    Schema schema = SchemaEncodeHelper.encode(
       SchemaType.UINT8, // 1 byte
       SchemaType.UINT16, // 2 bytes
       SchemaType.UINT32, // 4 bytes
@@ -232,7 +233,7 @@ contract SchemaTest is Test, GasReporter {
   }
 
   function testIsEmptyFalse() public {
-    Schema encodedSchema = SchemaLib.encode(SchemaType.UINT256);
+    Schema encodedSchema = SchemaEncodeHelper.encode(SchemaType.UINT256);
 
     startGasReport("check if schema is empty (non-empty schema)");
     bool empty = encodedSchema.isEmpty();