+++ templates/react/packages/client/src/App.tsx
@@ -1,11 +1,11 @@
 import { useComponentValue } from "@latticexyz/react";
 import { useMUD } from "./MUDContext";
+import { singletonEntity } from "@latticexyz/store-sync/recs";
 
 export const App = () => {
   const {
     components: { Counter },
     systemCalls: { increment },
-    network: { singletonEntity },
   } = useMUD();
 
   const counter = useComponentValue(Counter, singletonEntity);